Revert introduction of `forwardable`
Mongoid 7.1.0 introduced a regression (see
https://jira.mongodb.org/browse/MONGOID-4849) by switching to the
builtin ruby forwardable method of delegation rather than relying on the
ActiveSupport-provided delegation used before.

mongoid-locker was made compatible with the regressed version of mongoid
in https://github.com/mongodb/mongoid/pull/4739

Mongoid then fixed this regression in 7.1.1+ with
https://github.com/mongodb/mongoid/commit/d078acdfb0648e5b07ace16ee3077d14bf65b931,
which goes back to the active support way of delegating.

The `forwardable` inclusion in mongoid-locker causes incompatibilities
when you then include another gem that expects to be able to set up a
delegator within a mongoid document class. For example the
`mongoid-history` gem expects to be able to delegate using the
active-support style here:
https://github.com/mongoid/mongoid-history/blob/c8c4de1235c01252dfbdb9fe6c0abfc078ec1443/lib/mongoid/history/trackable.rb#L24-L25

This PR reverts the change to introduce `forwardable`.

See
https://github.com/mongoid/mongoid-history/issues/238#issuecomment-1063155193
for an example of how this interacts poorly with mongoid-history.